+++ lib/consult/cli.rb
@@ -38,7 +38,7 @@ module Consult
           opts[:force_render] = arg
         end
 
-        o.on '-v', '--quiet', FalseClass, 'Silence output' do |arg|
+        o.on '--quiet', FalseClass, 'Silence output' do |arg|
           opts[:verbose] = arg
         end
       end