+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### Unreleased
 
 * Stop with error when force rendering and no template is found
+* Remove non-functional short verbosity argument
 * Development: adjust require's to be able to run CLI from any folder
 
 #### 0.11.0