+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Stop with error when force rendering and no template is found
 * Remove non-functional short verbosity argument
+* Add `--version` option to show the version
 * Development: adjust require's to be able to run CLI from any folder
 
 #### 0.11.0